fix(css): move storybook css out of story file into local index.scss
