+++ packages/react/src/components/ButtonGroup/__stories__/ButtonGroup.stories.js
@@ -3,9 +3,10 @@ import { storiesOf } from '@storybook/react';
 import { withKnobs, text, select } from '@storybook/addon-knobs';
 import { ArrowRight20, ArrowDown20, Pdf20 } from '@carbon/icons-react';
 import '@carbon/grid/scss/grid.scss';
-import '../../../../../styles/scss/components/buttongroup/_buttongroup.scss';
 import readme from '../README.md';
 import { DDS_BUTTON_GROUP } from '../../../internal/FeatureFlags';
+
+import './index.scss';
 
 import ButtonGroup from '../ButtonGroup';
