+++ packages/react/src/components/ButtonGroup/__stories__/index.scss
@@ -1,0 +1,8 @@
+//
+// Copyright IBM Corp. 2016, 2018
+//
+// This source code is licensed under the Apache-2.0 license found in the
+// LICENSE file in the root directory of this source tree.
+//
+
+@import '../../../../../styles/scss/components/buttongroup/_buttongroup.scss';
