+++ packages/react/src/components/HorizontalRule/__stories__/HorizontalRule.stories.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, select, text } from '@storybook/addon-knobs';
-import '../../../../../styles/scss/components/horizontalrule/_horizontalrule.scss';
-import 'carbon-components/scss/globals/grid/_grid.scss';
 import readme from '../README.md';
+import './index.scss';
 
 import HorizontalRule from '../HorizontalRule';
