+++ packages/react/src/components/Layout/__stories__/index.scss
@@ -1,0 +1,9 @@
+//
+// Copyright IBM Corp. 2016, 2018
+//
+// This source code is licensed under the Apache-2.0 license found in the
+// LICENSE file in the root directory of this source tree.
+//
+
+@import '@carbon/grid/scss/grid.scss';
+@import '../../../../../styles/scss/components/layout/_layout.scss';
